refactor: extract CustomFieldValidator from CustomFieldRepository
Move validation logic (~300 lines) into a dedicated CustomFieldValidator
class: type-specific validation, format rules (CPF, email, phone, regex),
working_hours, and dependent_select. Repository keeps a delegating
validate_field_value() for backward compatibility. No caller changes
needed — all 66 existing tests pass.
